Remove LogAdapter, update AbstractReader#getParameters log level to DEBUG and update TestNG to 6.8.8
TestNG is not working, maybe because of my previous commit updating reflections to 9.10, not sure but updating to 6.8.8 make it work again.
The LogAdapter is doing nothing and I thinkg that AbstractReader#getParameter is in INFO level just because there are no DEBUG level in the adapter. I just removed, this actually broke the API because the LogAdapter was exposed at AbstractDocumentSource#getCustomApiReader which support classes with a (Swagger,LogAdapter) constructor. I think that this constructor is pretty obscure anyways so I removed it. I hope is ok.