Method arguments without annotation are now correctly handled for spring mvc
This implements the behaviour described in the springmvc doc:
"If a method argument is not matched to any of the above, by default it is resolved as an @RequestParam if it is a simple type, as determined by BeanUtils#isSimpleProperty, or as an @ModelAttribute otherwise."
https://docs.spring.io/spring/docs/current/spring-framework-reference/web.html#mvc-ann-arguments

Therefore, no annotations are needed for ModelAttribute and RequestParam.
Also, I fixed the behaviour of the defaultValue. The default was always set to an empty String. Now it is correctly taken from the annotation.
fixes #652