Support @DefaultValue for jax-rs
Correct the behavior of the DefaultValue annotation. It is now being applied to the Parameters.

This changes the extraction of the parameters for jaxrs, so that all of annotations are available at once. A contextual extraction is therefore now possible, allowing Annotations to access values of other annotations

This mostly matches how currently the SpringSwaggerExtension is implemented.